Do not return TEST_BAD when diagnostic is 1 (permanent system error). Fixes the second part of IR-338.
This probably needs some deeper thought. There seems to be some
uncertainty over whether returning TEST_BAD means "skip the rest of
this test" (which is what was intended here) or "skip this and all
sibling tests" (which is what's implemented). Do ANY instances of
TEST_BAD intend the latter? If not, we can just change the behaviour.